+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Section Order:
 ### Changed
 
 - Dashboard buttons to Font Awesome icons
+- Dashboard JS refactored
 
 ## [2.6.1] - 2025-08-19