+++ CHANGELOG.md
@@ -40,6 +40,7 @@ Section Order:
 - `views.ajax_dashboard` refactored
 - Dashboard buttons to Font Awesome icons
 - Dashboard JS refactored
+- Fleet details JS refactored
 
 ## [2.6.1] - 2025-08-19